Day Nine, part one

+++ day5/DayFive/AirConditioner.Tests/IntCodeCompilerTests.cs
@@ -14,8 +14,8 @@ namespace AirConditioner.Tests
         [DataRow(7, 0)]
         public void TestGiven1(int i, int e)
         {
-            IList<int> input = new List<int> { 3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8 };
-            var inputs = new Queue<int>();
+            IList<long> input = new List<long> { 3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8 };
+            var inputs = new Queue<long>();
             inputs.Enqueue(i);
             var icc = new IntCodeCompiler(input, inputs);
             icc.Calculate();
@@ -27,8 +27,8 @@ namespace AirConditioner.Tests
         [DataRow(7, 1)]
         public void TestGiven2(int i, int e)
         {
-            IList<int> input = new List<int> { 3, 9, 7, 9, 10, 9, 4, 9, 99, -1, 8 };
-            var inputs = new Queue<int>();
+            IList<long> input = new List<long> { 3, 9, 7, 9, 10, 9, 4, 9, 99, -1, 8 };
+            var inputs = new Queue<long>();
             inputs.Enqueue(i);
             var icc = new IntCodeCompiler(input, inputs);
             icc.Calculate();
@@ -41,8 +41,8 @@ namespace AirConditioner.Tests
         [DataRow(7, 0)]
         public void TestGiven3(int i, int e)
         {
-            IList<int> input = new List<int> { 3, 3, 1108, -1, 8, 3, 4, 3, 99 };
-            var inputs = new Queue<int>();
+            IList<long> input = new List<long> { 3, 3, 1108, -1, 8, 3, 4, 3, 99 };
+            var inputs = new Queue<long>();
             inputs.Enqueue(i);
             var icc = new IntCodeCompiler(input, inputs);
             icc.Calculate();
@@ -55,8 +55,8 @@ namespace AirConditioner.Tests
         [DataRow(7, 1)]
         public void TestGiven4(int i, int e)
         {
-            IList<int> input = new List<int> { 3, 3, 1107, -1, 8, 3, 4, 3, 99 };
-            var inputs = new Queue<int>();
+            IList<long> input = new List<long> { 3, 3, 1107, -1, 8, 3, 4, 3, 99 };
+            var inputs = new Queue<long>();
             inputs.Enqueue(i);
             var icc = new IntCodeCompiler(input, inputs);
             icc.Calculate();
@@ -68,8 +68,8 @@ namespace AirConditioner.Tests
         [DataRow(1, 1)]
         public void TestGiven5a(int i, int e)
         {
-            IList<int> input = new List<int> { 3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9 };
-            var inputs = new Queue<int>();
+            IList<long> input = new List<long> { 3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9 };
+            var inputs = new Queue<long>();
             inputs.Enqueue(i);
             var icc = new IntCodeCompiler(input, inputs);
             icc.Calculate();
@@ -80,8 +80,8 @@ namespace AirConditioner.Tests
         [DataRow(1, 1)]
         public void TestGiven5b(int i, int e)
         {
-            IList<int> input = new List<int> { 3, 3, 1105, -1, 9, 1101, 0, 0, 12, 4, 12, 99, 1 };
-            var inputs = new Queue<int>();
+            IList<long> input = new List<long> { 3, 3, 1105, -1, 9, 1101, 0, 0, 12, 4, 12, 99, 1 };
+            var inputs = new Queue<long>();
             inputs.Enqueue(i);
             var icc = new IntCodeCompiler(input, inputs);
             icc.Calculate();
@@ -93,8 +93,8 @@ namespace AirConditioner.Tests
         [DataRow(9, 1001)]
         public void TestGiven6(int i, int e)
         {
-            IList<int> input = new List<int> { 3, 21, 1008, 21, 8, 20, 1005, 20, 22, 107, 8, 21, 20, 1006, 20, 31, 1106, 0, 36, 98, 0, 0, 1002, 21, 125, 20, 4, 20, 1105, 1, 46, 104, 999, 1105, 1, 46, 1101, 1000, 1, 20, 4, 20, 1105, 1, 46, 98, 99 };
-            var inputs = new Queue<int>();
+            IList<long> input = new List<long> { 3, 21, 1008, 21, 8, 20, 1005, 20, 22, 107, 8, 21, 20, 1006, 20, 31, 1106, 0, 36, 98, 0, 0, 1002, 21, 125, 20, 4, 20, 1105, 1, 46, 104, 999, 1105, 1, 46, 1101, 1000, 1, 20, 4, 20, 1105, 1, 46, 98, 99 };
+            var inputs = new Queue<long>();
             inputs.Enqueue(i);
             var icc = new IntCodeCompiler(input, inputs);
             icc.Calculate();
